+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Added
 
 - Support to PMU events in Chapter 11
+- Support `NACL` extension in Chapter 15
 
 ### Modified